#1 Working client/server
