+++ g2diagnosticserver/g2diagnosticserver_test.go
@@ -1,0 +1,1 @@
+package g2diagnosticserver
